Use proper File.join() rather than home-grown slash concatenation.
It's a bad idea to hardcode system-specific path separators,
thus use the method that God intended to be used
for that purpose in Ruby (File.join()).

As a side effect, it's probably more efficient to pass string vars
as var arguments rather than doing manual string evaluation
on each implementation site.