Handle .obj files listed in VS10 AdditionalDependencies element.
For the somewhat dirty case of having external .obj files listed in
AdditionalDependencies,
the AdditionalDependencies content needs splitting and special treatment
since CMake expects object files to be supplied as source arguments
to a target.
Also, add a grave user-side syntax warning for this case,
since there might be better options.

Add a CMake conditional for the currently platform-specific .obj file handling,
and print a message on unsupported foreign platforms.