Bug fix: partially revert "Remove erroneous addition of MFC _AFXEXT, _AFXDLL defines."
This reverts commit 3c7eadfc0e0f9e63aee613b6b537c57fad10fb1a.

Turns out this seems to have been dead wrong.
On VS10 <UseOfMfc>Dynamic</UseOfMfc>, _AFXDLL definition *does* get added
implicitly, and there's lotsa breakage if we don't do likewise.
So perhaps it was something else that was quirky in that particular
case (ah, right, possibly we shouldn't be adding _AFXEXT - this is
something that seems to be added to the project defines explicitly,
for the Extension DLL case).
Thus restrict this setting to _AFXDLL only.