Bug fix: MIDL: VALIDATE_ALL_PARAMETERS actually is *not* an option-type param.
It's not a flag but a single-value param type.
This caused an incorrect call to the MIDL compilers, i.e. failure
to get things compiled.
I strongly suspect that this was uncovered by the recent change
to a local param parser implementation, since the value of
VALIDATE_ALL_PARAMETERS usually is "false", which is CMake-relevant syntax
content and thus probably was processed in a special manner by CMake's
CMakeParseArguments.cmake.