Bug fix: tempfile move handling was completely broken.
Tempfile move happened outside of file-*removing* block scope.
I really don't know how I managed to miss this when implementing it
(it's obvious).

Got uncovered due to sometimes causing exceptions on a heavily multi-core host.

And to add insult to injury, the code failed to do a *manual*(!!) close()
of the tempfile prior to moving it,
to have it properly ensured that the file is properly flushed once moving it...

** A R G H **